beam-sql-to-postgres/app/requirements.txt
apache-beam
psycopg2-binary
python-dotenv